+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Editor
   - Fix `didChangeWatchedFiles` to correctly create the file on server, properly change file content and re-scan with clj-kondo, or remove file analysis. This should improve LSP analysis reliability when changing files outside the editor. #536
+  - Improve completion only showing valid local vars for current cursor.
   
 - API/CLI
   - New `diagnostics` command, which return all diagnostics found by clojure-lsp (using clj-kondo). Check the API section for more details.